+++ scanapi/hide_utils.py
@@ -1,10 +1,11 @@
+import json
+
 from scanapi.settings import settings
 
 HEADERS = "headers"
 BODY = "body"
 URL = "url"
 
-ALLOWED_ATTRS_TO_HIDE = (HEADERS, BODY, URL)
 SENSITIVE_INFO_SUBSTITUTION_FLAG = "SENSITIVE_INFORMATION"
 
 
@@ -31,16 +32,28 @@ def _hide(http_msg, hide_settings):
 
 def _override_info(http_msg, http_attr, secret_field):
     """ Private method that substitutes sensitive data with string 'SENSITIVE_INFORMATION' """
-    if (
-        secret_field in getattr(http_msg, http_attr)
-        and http_attr in ALLOWED_ATTRS_TO_HIDE
-    ):
-        if http_attr == URL:
-            new_url = getattr(http_msg, http_attr).replace(
-                secret_field, SENSITIVE_INFO_SUBSTITUTION_FLAG
-            )
-            setattr(http_msg, http_attr, new_url)
-        else:
-            getattr(http_msg, http_attr)[
-                secret_field
-            ] = SENSITIVE_INFO_SUBSTITUTION_FLAG
+
+    if http_attr == URL:
+        _override_url(http_msg, secret_field)
+    elif http_attr == HEADERS:
+        _override_headers(http_msg, secret_field)
+    elif http_attr == BODY:
+        _override_body(http_msg, secret_field)
+
+
+def _override_url(http_msg, secret_field):
+    if secret_field in http_msg.url:
+        new_url = http_msg.url.replace(secret_field, SENSITIVE_INFO_SUBSTITUTION_FLAG)
+        http_msg.url = new_url
+
+
+def _override_headers(http_msg, secret_field):
+    if secret_field in http_msg.headers:
+        http_msg.headers[secret_field] = SENSITIVE_INFO_SUBSTITUTION_FLAG
+
+
+def _override_body(http_msg, secret_field):
+    body = json.loads(http_msg.body.decode("UTF-8"))
+    if secret_field in body:
+        body[secret_field] = SENSITIVE_INFO_SUBSTITUTION_FLAG
+        http_msg.body = json.dumps(body).encode("utf-8")